Revert "Fix tests"
This reverts commit 093d922de59e5908872197ab40befceda9af11b6.
